feat: Badge preview + Challenges (#120)
* Keep actions if the wallet is not connected

* Improve badge listed on PendingList
Add navigation to badge preview

* fix logout

---------
