+++ src/components/header/UserDropdown.tsx
@@ -1,11 +1,9 @@
 import { useRouter } from 'next/router'
-import React, { RefObject, useState } from 'react'
+import React, { useState } from 'react'
 
 import {
   Avatar,
   Badge,
-  Box,
-  Button,
   Divider,
   IconButton,
   ListItemIcon,
@@ -14,16 +12,13 @@ import {
   Tooltip,
   styled,
 } from '@mui/material'
-import { useTranslation } from 'next-export-i18n'
 import Blockies from 'react-18-blockies'
-import { colors } from 'thebadge-ui-library'
 
 import { Logout } from '@/src/components/assets/Logout'
 import { SwitchNetwork } from '@/src/components/assets/SwitchNetwork'
+import ActionButtons from '@/src/components/header/ActionButtons'
 import { ModalSwitchNetwork } from '@/src/components/helpers/ModalSwitchNetwork'
 import SafeSuspense from '@/src/components/helpers/SafeSuspense'
-import { useCurrentUser } from '@/src/hooks/subgraph/useCurrentUser'
-import { useSectionReferences } from '@/src/providers/referencesProvider'
 import { useWeb3Connection } from '@/src/providers/web3ConnectionProvider'
 import { truncateStringInTheMiddle } from '@/src/utils/strings'
 
@@ -56,24 +51,9 @@ const StyledBadge = styled(Badge)<{ state?: 'ok' | 'error' }>(({ state, theme })
   },
 }))
 
-const StyledButton = styled(Button)<{ border?: string }>(({ border }) => ({
-  color: 'white',
-  border,
-  borderRadius: '10px',
-  fontSize: '14px !important',
-  padding: '0.5rem 1rem !important',
-  height: 'fit-content !important',
-  lineHeight: '14px',
-  fontWeight: 700,
-  boxShadow: 'none',
-}))
-
 export const UserDropdown: React.FC = () => {
   const router = useRouter()
-  const { t } = useTranslation()
   const { address, disconnectWallet, isWalletNetworkSupported } = useWeb3Connection()
-  const { data: user } = useCurrentUser()
-  const { becomeACreatorSection, scrollTo } = useSectionReferences()
 
   const [showNetworkModal, setShowNetworkModal] = useState(false)
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null)
@@ -85,38 +65,10 @@ export const UserDropdown: React.FC = () => {
     setAnchorEl(null)
   }
 
-  const menuButton = (
-    title: string,
-    color: string,
-    disabled: boolean,
-    path: string,
-    ref?: RefObject<HTMLDivElement> | null,
-  ) => {
-    return (
-      <StyledButton
-        border={`2px solid ${color}`}
-        disabled={disabled}
-        onClick={() => {
-          scrollTo(path, ref || null)
-        }}
-      >
-        {title}
-      </StyledButton>
-    )
+  const logout = async () => {
+    await disconnectWallet()
+    router.push('/')
   }
-  const exploreButton = menuButton(t('header.buttons.mint'), colors.blue, false, '/badge/explorer')
-  const curateButton = menuButton(
-    t('header.buttons.curate'),
-    colors.greenLogo,
-    false,
-    '/badge/curate',
-  )
-  const createButton = menuButton(
-    t('header.buttons.create'),
-    colors.pink,
-    !user || !user?.isCreator,
-    '/badge/model/create',
-  )
 
   const blockiesIcon = (
     <SafeSuspense>
@@ -126,37 +78,7 @@ export const UserDropdown: React.FC = () => {
 
   return (
     <>
-      <Box
-        alignItems="center"
-        display="flex"
-        flex={1}
-        justifyContent="space-between"
-        sx={{ columnGap: '10px' }}
-      >
-        {exploreButton}
-        {curateButton}
-
-        {!user || !user?.isCreator ? (
-          <Tooltip
-            arrow
-            title={
-              <>
-                {t('header.tooltips.becomeACreator.prefixText')}
-                <Box
-                  onClick={() => scrollTo('/', becomeACreatorSection)}
-                  style={{ cursor: 'pointer', textDecoration: 'underline' }}
-                >
-                  {t('header.tooltips.becomeACreator.link')}
-                </Box>
-              </>
-            }
-          >
-            <span>{createButton}</span>
-          </Tooltip>
-        ) : (
-          createButton
-        )}
-      </Box>
+      <ActionButtons />
       <Tooltip arrow title="Account settings">
         <IconButton
           aria-controls={open ? 'account-menu' : undefined}
@@ -224,7 +146,7 @@ export const UserDropdown: React.FC = () => {
           </ListItemIcon>
           Switch network
         </MenuItem>
-        <MenuItem onClick={disconnectWallet}>
+        <MenuItem onClick={logout}>
           <ListItemIcon>
             <Logout />
           </ListItemIcon>
