+++ src/pagePartials/badge/BadgeModelPreview.tsx
@@ -18,7 +18,7 @@ function BadgeModelPreview({ metadata, size = 'medium' }: Props) {
   return (
     <SafeSuspense>
       <BadgePreview
-        animationEffects={[]}
+        animationEffects={['wobble', 'grow', 'glare']}
         animationOnHover
         badgeBackgroundUrl="https://images.unsplash.com/photo-1512998844734-cd2cca565822?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxzZWFyY2h8MTIyfHxhYnN0cmFjdHxlbnwwfHwwfHw%3D&auto=format&fit=crop&w=500&q=60"
         badgeUrl="https://www.thebadge.xyz"
