+++ src/pagePartials/index/Head.tsx
@@ -9,7 +9,7 @@ export const Head = () => {
   const siteURL = typeof window !== 'undefined' ? `${protocol}//${hostname}${portString}` : ''
   const title = 'TheBadge - DApp'
   const description = 'TheBadge - DApp'
-  const twitterHandle = '@'
+  const twitterHandle = '@thebadgexyz'
 
   return (
     <NextHead>
@@ -17,7 +17,7 @@ export const Head = () => {
       <meta content={description} name="description" />
       <meta content={title} property="og:title" />
       <meta content={siteURL} property="og:url" />
-      <meta content={`${siteURL}/shareable/ogImage.jpg`} property="og:image" />
+      <meta content={`${siteURL}/shareable/the_badge_banner.webp`} property="og:image" />
       <meta content="website" property="og:type" />
       <meta content={description} property="og:description" />
       <meta content="summary_large_image" name="twitter:card" />
